Created interface for UDP Socket
