+++ src/udp_unix.cpp
@@ -1,0 +1,61 @@
+#include "udp.h"
+
+namespace udp {
+
+    Socket::Socket(const std::string& ipAddr, uint16_t port, bool reciever) {
+        sIPAddr = ipAddr;
+        unPort = port;
+        bReceiver = reciever;
+
+        memset(&stSenderAddress, 0, sizeof(sockaddr_in));
+        memset(&stRecieverAddress, 0, sizeof(sockaddr_in));
+
+        stRecieverAddress.sin_family = AF_INET;
+        stRecieverAddress.sin_port = htons(port);
+        stRecieverAddress.sin_addr.s_addr = inet_addr(ipAddr.c_str());
+    }
+
+    bool StartSocket(Socket& stSocket) {
+        int nFileDescriptor = socket(AF_INET, SOCK_DGRAM, 0);
+        if(nFileDescriptor < 0) return false;
+        else stSocket.nFileDescriptor = nFileDescriptor;
+
+        if(stSocket.bReceiver) {
+            int nResult = bind(stSocket.nFileDescriptor, (const sockaddr*) &stSocket.stRecieverAddress, sizeof(sockaddr));
+            if(nResult < 0) return false;
+        }
+
+        return true;
+    }
+
+    void CloseSocket(Socket& stSocket) {
+        close(stSocket.nFileDescriptor);
+    }
+
+    bool Send(Socket& stSocket, std::vector<unsigned char>& sMessage) {
+        int nResult = sendto(stSocket.nFileDescriptor,
+               sMessage.data(),
+               sMessage.size(),
+               MSG_CONFIRM,
+               (const sockaddr *) &stSocket.stRecieverAddress,
+               sizeof(stSocket.stRecieverAddress));
+        return nResult >= 0 ? true : false;
+    }
+
+    void Receive(Socket& stSocket, std::vector<unsigned char>& sMessage, uint64_t unBufferSize) {
+        // char aBuffer[unBufferSize];
+        unsigned char* pBuffer = new unsigned char[unBufferSize];
+        socklen_t siAddrLen = sizeof(stSocket.stSenderAddress);
+        int nMessageLength = recvfrom(stSocket.nFileDescriptor, 
+                // (char *) aBuffer,
+                pBuffer, 
+                unBufferSize, 
+                MSG_WAITALL, 
+                (sockaddr *) &stSocket.stSenderAddress, 
+                &siAddrLen);
+        // aBuffer[nMessageLength] = '\0';
+        // sMessage = std::string(aBuffer);
+        sMessage = std::vector<unsigned char>(pBuffer, pBuffer + nMessageLength);
+    }
+
+};
